+++ source/G3Dcpp/MeshAlgWeld.cpp
@@ -87,7 +87,7 @@ Welder::Welder(
     Vector3 minBound = Vector3::inf();
     Vector3 maxBound = -minBound;
 
-    for (int i = 0; i < oldVertexArray.size(); ++i) {
+    for (size_t i = 0; i < oldVertexArray.size(); ++i) {
         minBound.min(oldVertexArray[i]);
         maxBound.max(oldVertexArray[i]);
     }
@@ -124,7 +124,7 @@ int Welder::getIndex(const Vector3& vertex) {
     // Check against all vertices within radius of this grid cube
     const List& list = grid[ix][iy][iz];
 
-    for (int i = 0; i < list.size(); ++i) {
+    for (size_t i = 0; i < list.size(); ++i) {
         double d = (newVertexArray[list[i]] - vertex).squaredLength();
 
         if (d < distanceSquared) {
@@ -175,7 +175,7 @@ void Welder::weld() {
     newVertexArray.resize(0);
 
     // Prime the vertex positions
-    for (int i = 0; i < oldVertexArray.size(); ++i) {
+    for (size_t i = 0; i < oldVertexArray.size(); ++i) {
         getIndex(oldVertexArray[i]);
     }
 
@@ -184,7 +184,7 @@ void Welder::weld() {
     toNew.resize(oldVertexArray.size());
     toOld.resize(newVertexArray.size());
 
-    for (int oi = 0; oi < oldVertexArray.size(); ++oi) {
+    for (size_t oi = 0; oi < oldVertexArray.size(); ++oi) {
         toNew[oi] = getIndex(oldVertexArray[oi]);
         toOld[toNew[oi]] = oi;
     }
